+++ src/java/Model/TriangleAreaService.java
@@ -6,8 +6,8 @@ package Model;
  */
 public class TriangleAreaService {
     public double getArea(String userBase,String userHeight){
-        int base=userBase!=null?Integer.parseInt(userBase):0;
-        int height=userHeight!=null?Integer.parseInt(userHeight):0;
+        double base=userBase!=null?Double.parseDouble(userBase):0;
+        double height=userHeight!=null?Double.parseDouble(userHeight):0;
         
         return (base*height)/2;
     }
